Add the option of using reattach-to-user-namespace
When OS X runs as a scheduled service in OS X, it runs as root in a way
that shelling out, even as a specific user, results in bad responses
from Mas. The reattach-to-user-namespace utility can help with this, but
let's leave it disabled by default.

```
$ mas account
j@p4nt5.com

$ sudo mas account
Not signed in

$ sudo su -
$ mas account
Not signed in

$ sudo -u jhartman mas account
Not signed in

$ sudo -u jhartman reattach-to-user-namespace mas account
j@p4nt5.com
```